requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.5
Deprecated==1.2.14
Django==5.0.3
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
frozenlist==1.4.1
idna==3.6
kombu==5.3.5
multidict==6.0.5
prompt-toolkit==3.0.43
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
twilio==8.13.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
vonage==3.13.0
vonage-jwt==1.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
